feat: Headers, footers, some styling
