+++ src/consts.ts
@@ -1,5 +1,6 @@
 // Place any global data in this file.
 // You can import this data from anywhere in your site by using the `import` keyword.
 
-export const SITE_TITLE = 'Dylan Anthony';
-export const SITE_DESCRIPTION = 'Producing quality software with minimal frustration.';
+export const SITE_TITLE = "Dylan Anthony";
+export const SITE_DESCRIPTION =
+  "Producing quality software with minimal frustration.";
