+++ src/content/blog/using-mdx.mdx
@@ -1,8 +1,8 @@
 ---
-title: 'Using MDX'
-description: 'Lorem ipsum dolor sit amet'
-pubDate: 'Jul 02 2022'
-heroImage: '/blog-placeholder-5.jpg'
+title: "Using MDX"
+description: "Lorem ipsum dolor sit amet"
+pubDate: "Jul 02 2022"
+heroImage: "/blog-placeholder-5.jpg"
 ---
 
 This theme comes with the [@astrojs/mdx](https://docs.astro.build/en/guides/integrations-guide/mdx/) integration installed and configured in your `astro.config.mjs` config file. If you prefer not to use MDX, you can disable support by removing the integration from your config file.
@@ -18,10 +18,10 @@ If you have existing content authored in MDX, this integration will hopefully ma
 Here is how you import and use a UI component inside of MDX.  
 When you open this page in the browser, you should see the clickable button below.
 
-import HeaderLink from '../../components/HeaderLink.astro';
+import HeaderLink from "../../components/HeaderLink.astro";
 
 <HeaderLink href="#" onclick="alert('clicked!')">
-	Embedded component in MDX
+  Embedded component in MDX
 </HeaderLink>
 
 ## More Links
