+++ src/layouts/Body.astro
@@ -1,0 +1,12 @@
+---
+
+import Header from "../components/Header.astro";
+import Footer from "../components/Footer.astro";---
+
+<body class="bg-indigo-50 dark:bg-zinc-800 text-zinc-700 dark:text-zinc-200">
+<div class="container mx-auto px-2 md:px-20 pt-5 mb-6">
+    <Header/>
+    <slot/>
+    <Footer/>
+</div>
+</body>
